pyproject.toml
[tool.poetry]
name = "instrumentdb"
version = "0.1.0"
description = ""
authors = ["Maurizio Tomasi <ziotom78@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.3"
markdown = "^3.2.1"
pygments = "^2.5.2"
djangorestframework = "^3.11.0"
django-filter = "^2.2.0"
uuid = "^1.30"
pyyaml = "^5.3"
django-mptt = "^0.11.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
manage = "manage:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
